+++ src/outputs/outputs.cpp
@@ -57,17 +57,20 @@
 
 #include "outputs/outputs.hpp"
 
+#include <algorithm>
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
 #include <iomanip>
 #include <iostream>
+#include <set>
 #include <sstream>
 #include <stdexcept>
 #include <string>
 
 #include "coordinates/coordinates.hpp"
 #include "defs.hpp"
+#include "globals.hpp"
 #include "mesh/mesh.hpp"
 #include "mesh/meshblock.hpp"
 #include "parameter_input.hpp"
@@ -75,47 +78,6 @@
 #include "utils/error_checking.hpp"
 
 namespace parthenon {
-
-//----------------------------------------------------------------------------------------
-// VarInfo constructor
-VarInfo::VarInfo(const std::string &label,
-                 const std::vector<std::string> &component_labels_, int num_components,
-                 int nx6, int nx5, int nx4, int nx3, int nx2, int nx1, Metadata metadata,
-                 bool is_sparse, bool is_vector)
-    : label(label), num_components(num_components), nx6(nx6), nx5(nx5), nx4(nx4),
-      nx3(nx3), nx2(nx2), nx1(nx1), tensor_rank(metadata.Shape().size()),
-      where(metadata.Where()), is_sparse(is_sparse), is_vector(is_vector) {
-  if (num_components <= 0) {
-    std::stringstream msg;
-    msg << "### ERROR: Got variable " << label << " with " << num_components
-        << " components."
-        << " num_components must be greater than 0" << std::endl;
-    PARTHENON_FAIL(msg);
-  }
-
-  // Note that this logic does not subscript components without component_labels if
-  // there is only one component. Component names will be e.g.
-  //   my_scalar
-  // or
-  //   my_non-vector_set_0
-  //   my_non-vector_set_1
-  // Note that this means the subscript will be dropped for multidim quantities if their
-  // Nx6, Nx5, Nx4 are set to 1 at runtime e.g.
-  //   my_non-vector_set
-  // Similarly, if component labels are given for all components, those will be used
-  // without the prefixed label.
-  component_labels = {};
-  if (num_components == 1 || is_vector) {
-    component_labels = component_labels_.size() > 0 ? component_labels_
-                                                    : std::vector<std::string>({label});
-  } else if (component_labels_.size() == num_components) {
-    component_labels = component_labels_;
-  } else {
-    for (int i = 0; i < num_components; i++) {
-      component_labels.push_back(label + "_" + std::to_string(i));
-    }
-  }
-}
 
 //----------------------------------------------------------------------------------------
 // OutputType constructor
@@ -195,9 +157,9 @@ Outputs::Outputs(Mesh *pm, ParameterInput *pin, SimTime *tm) {
 
         if (pin->DoesParameterExist(op.block_name, "single_precision_output")) {
           std::stringstream warn;
-          warn << "### WARNING Output option single_precision_output only applies to "
+          warn << "Output option single_precision_output only applies to "
                   "HDF5 outputs or restarts. Ignoring it for output block '"
-               << op.block_name << "'" << std::endl;
+               << op.block_name << "'";
           PARTHENON_WARN(warn);
         }
       }
@@ -214,8 +176,8 @@ Outputs::Outputs(Mesh *pm, ParameterInput *pin, SimTime *tm) {
 #ifdef PARTHENON_DISABLE_HDF5_COMPRESSION
         if (op.hdf5_compression_level != 0) {
           std::stringstream err;
-          err << "### ERROR: HDF5 compression requested for output block '"
-              << op.block_name << "', but HDF5 compression is disabled" << std::endl;
+          err << "HDF5 compression requested for output block '" << op.block_name
+              << "', but HDF5 compression is disabled";
           PARTHENON_THROW(err)
         }
 #endif
@@ -224,9 +186,9 @@ Outputs::Outputs(Mesh *pm, ParameterInput *pin, SimTime *tm) {
 
         if (pin->DoesParameterExist(op.block_name, "hdf5_compression_level")) {
           std::stringstream warn;
-          warn << "### WARNING Output option hdf5_compression_level only applies to "
+          warn << "Output option hdf5_compression_level only applies to "
                   "HDF5 outputs or restarts. Ignoring it for output block '"
-               << op.block_name << "'" << std::endl;
+               << op.block_name << "'";
           PARTHENON_WARN(warn);
         }
       }
@@ -234,7 +196,40 @@ Outputs::Outputs(Mesh *pm, ParameterInput *pin, SimTime *tm) {
       // set output variable and optional data format string used in formatted writes
       if ((op.file_type != "hst") && (op.file_type != "rst") &&
           (op.file_type != "ascent")) {
-        op.variables = pin->GetVector<std::string>(pib->block_name, "variables");
+        op.variables = pin->GetOrAddVector<std::string>(pib->block_name, "variables",
+                                                        std::vector<std::string>());
+        // JMM: If the requested var isn't present for a given swarm,
+        // it is simply not output.
+        op.swarms.clear(); // Not sure this is needed
+        if (pin->DoesParameterExist(pib->block_name, "swarms")) {
+          std::vector<std::string> swarmnames =
+              pin->GetVector<std::string>(pib->block_name, "swarms");
+          std::size_t nswarms = swarmnames.size();
+          if ((pin->DoesParameterExist(pib->block_name, "swarm_variables")) &&
+              (nswarms > 1)) {
+            std::stringstream msg;
+            msg << "The swarm_variables field is set in the block '" << pib->block_name
+                << "' however, there are " << nswarms << " swarms."
+                << " All swarms will be assumed to request the vars listed in "
+                   "swarm_variables.";
+            PARTHENON_WARN(msg);
+          }
+          for (const auto &swname : swarmnames) {
+            if (pin->DoesParameterExist(pib->block_name, "swarm_variables")) {
+              auto varnames =
+                  pin->GetVector<std::string>(pib->block_name, "swarm_variables");
+              op.swarms[swname].insert(varnames.begin(), varnames.end());
+            }
+            if (pin->DoesParameterExist(pib->block_name, swname + "_variables")) {
+              auto varnames =
+                  pin->GetVector<std::string>(pib->block_name, swname + "_variables");
+              op.swarms[swname].insert(varnames.begin(), varnames.end());
+            }
+            // Always output x, y, and z for swarms so that they work with vis tools.
+            std::vector<std::string> coords = {"x", "y", "z"};
+            op.swarms[swname].insert(coords.begin(), coords.end());
+          }
+        }
       }
       op.data_format = pin->GetOrAddString(op.block_name, "data_format", "%12.5e");
       op.data_format.insert(0, " "); // prepend with blank to separate columns